Move tls testing functions to tlstest
So that the testing logic not build in release binary
