+++ internal/server/endpoints/endpoints_test.go
@@ -20,6 +20,7 @@ import (
 	localUtil "github.com/lxc/incus/v6/internal/server/util"
 	"github.com/lxc/incus/v6/shared/api"
 	localtls "github.com/lxc/incus/v6/shared/tls"
+	"github.com/lxc/incus/v6/shared/tls/tlstest"
 	"github.com/lxc/incus/v6/shared/util"
 )
 
@@ -37,7 +38,7 @@ func newEndpoints(t *testing.T) (*endpoints.Endpoints, *endpoints.Config, func()
 		UnixSocket:     filepath.Join(dir, "unix.socket"),
 		RestServer:     newServer(),
 		DevIncusServer: newServer(),
-		Cert:           localtls.TestingKeyPair(),
+		Cert:           tlstest.TestingKeyPair(t),
 		VsockServer:    newServer(),
 	}
